feat: split localization result
